Enable multi-process for API service
Due to the limit of Python interpreter, API service of Manila can't
really utilize underlying multi-core architecture even libraries
like eventlet has been used. To make API service much more scalable,
we'd adopt multi-process (worker) mode that has been used for long
in Glance/Swift/Nova/Cinder.

The default behavior isn't changed with this patch, Manila API
service will still run in one process (default value of
osapi_volume_workers is None).

Implementation wise, a good portion of manila/service.py has been
removed because those content has been merged in Oslo version of
service module.  manila/wsgi.py is also updated to adopt the change
for multiple WSGI servers running in separate processes.

DocImpact: 'New config option osapi_share_workers is used to specify
number of API service workers (OS processes) to launch for Manila
API service.  Setting this config option to a proper value (e.g.
osapi_share_workers = # of CPU cores/threads of the machine) can
greatly improve the total throughput of API service [# of API
requests can be handled per second].'

Also removed out-dated comments in manila/cmd/api.py due to the fact
that this bug [1] has been fixed in eventlet 0.9.13

This commit is port from two Cinder changes - [2] and [3].

[1] https://bitbucket.org/eventlet/eventlet/issue/92/eventletgreen-override-of-oswaitpid
[2] I8361d0dc0d43040e48634ff1aee1324e5e0af466
[3] Iab32a3fe230a11692a8cad274304214247d6c2c6
Implement bp: multi-process-api-service

+++ manila/wsgi.py
@@ -31,7 +31,8 @@ import eventlet.wsgi
 import greenlet
 from oslo_config import cfg
 from oslo_log import log
-from oslo_utils import netutils
+from oslo_service import service
+from oslo_utils import excutils
 from paste import deploy
 import routes.middleware
 import six
@@ -40,6 +41,7 @@ import webob.exc
 
 from manila import exception
 from manila.i18n import _
+from manila.i18n import _LE
 from manila.i18n import _LI
 
 socket_opts = [
@@ -91,13 +93,13 @@ CONF.register_opts(eventlet_opts)
 LOG = log.getLogger(__name__)
 
 
-class Server(object):
+class Server(service.ServiceBase):
     """Server class to manage a WSGI server, serving a WSGI application."""
 
     default_pool_size = 1000
 
     def __init__(self, name, app, host=None, port=None, pool_size=None,
-                 protocol=eventlet.wsgi.HttpProtocol):
+                 protocol=eventlet.wsgi.HttpProtocol, backlog=128):
         """Initialize, but do not start, a WSGI server.
 
         :param name: Pretty name for logging.
@@ -116,10 +118,14 @@ class Server(object):
         self._server = None
         self._socket = None
         self._protocol = protocol
-        self._pool = eventlet.GreenPool(pool_size or self.default_pool_size)
+        self.pool_size = pool_size or self.default_pool_size
+        self._pool = eventlet.GreenPool(self.pool_size)
         self._logger = log.getLogger("eventlet.wsgi.server")
 
-    def _get_socket(self, host, port, backlog):
+        if backlog < 1:
+            raise exception.InvalidInput(
+                reason='The backlog must be more than 1')
+
         bind_addr = (host, port)
         # TODO(dims): eventlet's green dns/socket module does not actually
         # support IPv6 in getaddrinfo(). We need to get around this in the
@@ -137,7 +143,7 @@ class Server(object):
         cert_file = CONF.ssl_cert_file
         key_file = CONF.ssl_key_file
         ca_file = CONF.ssl_ca_file
-        use_ssl = cert_file or key_file
+        self._use_ssl = cert_file or key_file
 
         if cert_file and not os.path.exists(cert_file):
             raise RuntimeError(_("Unable to find cert_file : %s") % cert_file)
@@ -148,85 +154,85 @@ class Server(object):
         if key_file and not os.path.exists(key_file):
             raise RuntimeError(_("Unable to find key_file : %s") % key_file)
 
-        if use_ssl and (not cert_file or not key_file):
+        if self._use_ssl and (not cert_file or not key_file):
             raise RuntimeError(_("When running server in SSL mode, you must "
                                  "specify both a cert_file and key_file "
                                  "option value in your configuration file"))
 
-        def wrap_ssl(sock):
-            ssl_kwargs = {
-                'server_side': True,
-                'certfile': cert_file,
-                'keyfile': key_file,
-                'cert_reqs': ssl.CERT_NONE,
-            }
-
-            if CONF.ssl_ca_file:
-                ssl_kwargs['ca_certs'] = ca_file
-                ssl_kwargs['cert_reqs'] = ssl.CERT_REQUIRED
-
-            return ssl.wrap_socket(sock, **ssl_kwargs)
-
-        sock = None
         retry_until = time.time() + 30
-        while not sock and time.time() < retry_until:
+        while not self._socket and time.time() < retry_until:
             try:
-                sock = eventlet.listen(bind_addr,
-                                       backlog=backlog,
-                                       family=family)
-                if use_ssl:
-                    sock = wrap_ssl(sock)
-
+                self._socket = eventlet.listen(
+                    bind_addr, backlog=backlog, family=family)
             except socket.error as err:
                 if err.args[0] != errno.EADDRINUSE:
                     raise
                 eventlet.sleep(0.1)
-        if not sock:
+
+        if not self._socket:
             raise RuntimeError(_("Could not bind to %(host)s:%(port)s "
                                "after trying for 30 seconds") %
                                {'host': host, 'port': port})
-        sock.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
 
-        # sockets can hang around forever without keepalive
-        netutils.set_tcp_keepalive(sock,
-                                   CONF.tcp_keepalive,
-                                   CONF.tcp_keepidle,
-                                   CONF.tcp_keepalive_interval,
-                                   CONF.tcp_keepalive_count)
+        (self._host, self._port) = self._socket.getsockname()[0:2]
+        LOG.info(_LI("%(name)s listening on %(_host)s:%(_port)s"),
+                 {'name': self.name, '_host': self._host, '_port': self._port})
 
-        return sock
-
-    def _start(self):
-        """Run the blocking eventlet WSGI server.
-
-        :returns: None
-
-        """
-        eventlet.wsgi.server(self._socket,
-                             self.app,
-                             protocol=self._protocol,
-                             custom_pool=self._pool,
-                             log=self._logger)
-
-    def start(self, backlog=128):
+    def start(self):
         """Start serving a WSGI application.
 
-        :param backlog: Maximum number of queued connections.
         :returns: None
         :raises: manila.exception.InvalidInput
 
         """
-        if backlog < 1:
-            raise exception.InvalidInput(
-                reason='The backlog must be more than 1')
+        # The server socket object will be closed after server exits,
+        # but the underlying file descriptor will remain open, and will
+        # give bad file descriptor error. So duplicating the socket object,
+        # to keep file descriptor usable.
 
-        self._socket = self._get_socket(self._host,
-                                        self._port,
-                                        backlog=backlog)
-        self._server = eventlet.spawn(self._start)
-        (self._host, self._port) = self._socket.getsockname()[0:2]
-        LOG.info(_LI("Started %(name)s on %(_host)s:%(_port)s"),
-                 {'name': self.name, '_host': self._host, '_port': self._port})
+        dup_socket = self._socket.dup()
+        if self._use_ssl:
+            try:
+                ssl_kwargs = {
+                    'server_side': True,
+                    'certfile': CONF.ssl_cert_file,
+                    'keyfile': CONF.ssl_key_file,
+                    'cert_reqs': ssl.CERT_NONE,
+                }
+
+                if CONF.ssl_ca_file:
+                    ssl_kwargs['ca_certs'] = CONF.ssl_ca_file
+                    ssl_kwargs['cert_reqs'] = ssl.CERT_REQUIRED
+
+                dup_socket = ssl.wrap_socket(dup_socket,
+                                             **ssl_kwargs)
+
+                dup_socket.setsockopt(socket.SOL_SOCKET,
+                                      socket.SO_REUSEADDR, 1)
+
+                # sockets can hang around forever without keepalive
+                dup_socket.setsockopt(socket.SOL_SOCKET,
+                                      socket.SO_KEEPALIVE, 1)
+
+            except Exception:
+                with excutils.save_and_reraise_exception():
+                    LOG.error(
+                        _LE("Failed to start %(name)s on %(_host)s:%(_port)s "
+                            "with SSL support."),
+                        {"name": self.name, "_host": self._host,
+                         "_port": self._port}
+                    )
+
+        wsgi_kwargs = {
+            'func': eventlet.wsgi.server,
+            'sock': dup_socket,
+            'site': self.app,
+            'protocol': self._protocol,
+            'custom_pool': self._pool,
+            'log': self._logger,
+        }
+
+        self._server = eventlet.spawn(**wsgi_kwargs)
 
     @property
     def host(self):
@@ -246,7 +252,10 @@ class Server(object):
 
         """
         LOG.info(_LI("Stopping WSGI server."))
-        self._server.kill()
+        if self._server is not None:
+            # Resize pool to stop new requests from being processed
+            self._pool.resize(0)
+            self._server.kill()
 
     def wait(self):
         """Block, until the server has stopped.
@@ -257,9 +266,18 @@ class Server(object):
 
         """
         try:
-            self._server.wait()
+            if self._server is not None:
+                self._pool.waitall()
+                self._server.wait()
         except greenlet.GreenletExit:
             LOG.info(_LI("WSGI server has stopped."))
+
+    def reset(self):
+        """Reset server greenpool size to default.
+
+        :returns: None
+        """
+        self._pool.resize(self.pool_size)
 
 
 class Request(webob.Request):
